♻️ Move paging MVP into amd64 module
